requirements.txt
beautifulsoup4
python-dotenv
pandas
selenium
webdriver_manager
spacy
psycopg2-binary
pytest
pylint
rapidfuzz
streamlit
nltk
streamlit_folium
ipykernel
plotly_express